requirements.txt
streamlit>=1.35
ultralytics>=8.2
opencv-python-headless>=4.8
pillow>=10.0
numpy>=1.26
